+++ bespokeasm/assembler/pretty_printer/factory.py
@@ -1,6 +1,7 @@
 from bespokeasm.assembler.pretty_printer import PrettyPrinterBase
 from bespokeasm.assembler.pretty_printer.source_details import SourceDetailsPrettyPrinter
 from bespokeasm.assembler.pretty_printer.minhex import MinHexPrettyPrinter
+from bespokeasm.assembler.pretty_printer.intelhex import IntelHexPrettyPrinter
 from bespokeasm.assembler.line_object import LineObject
 from bespokeasm.assembler.model import AssemblerModel
 
@@ -18,4 +19,8 @@ class PrettyPrinterFactory:
             return SourceDetailsPrettyPrinter(line_objs, model)
         elif pretty_printer_type == 'minhex':
             return MinHexPrettyPrinter(line_objs, model)
+        elif pretty_printer_type == 'hex':
+            return IntelHexPrettyPrinter(line_objs, model, False)
+        elif pretty_printer_type == 'intel_hex':
+            return IntelHexPrettyPrinter(line_objs, model, True)
         raise NotImplementedError